requirements.txt
asgiref==3.4.1
black==21.9b0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==3.9.2
idna==3.2
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
platformdirs==2.3.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
